updated calculations

+++ src/user/deposit/deposit.service.ts
@@ -70,11 +70,11 @@ export class DepositService {
 
     async getInvestorProAmountByUser(userId: number): Promise<InvestorProDepositAmountReponse> {
         const [{sum: allPackages}] = (await this.depositRepo.query(
-            `select cast(coalesce(sum(d.currency_amount),0)as int) as sum from "public".deposit d where d.product_service_description like 'investor_pro%' and d.product_service_description != 'investor_pro_gamefi'`,
+            `select cast(coalesce(sum(d.currency_amount),0)as float) as sum from "public".deposit d where d.product_service_description like 'investor_pro%' and d.product_service_description != 'investor_pro_gamefi'`,
         )) as [{ sum: number }];
 
         const [{sum: perUser}] = (await this.depositRepo.query(
-            `select cast(coalesce(sum(d.currency_amount),0)as int) as sum from "public".deposit d where d.product_service_description like 'investor_pro%' and d.product_service_description != 'investor_pro_gamefi' and d."userId" = $1`,
+            `select cast(coalesce(sum(d.currency_amount),0)as float) as sum from "public".deposit d where d.product_service_description like 'investor_pro%' and d.product_service_description != 'investor_pro_gamefi' and d."userId" = $1`,
             [userId],
         )) as [{ sum: number }];
 
@@ -90,19 +90,19 @@ export class DepositService {
         }] = await this.depositRepo.query(
             `
             select (
-              select cast(coalesce(sum(d.currency_amount), 0) as int)
+              select cast(coalesce(sum(d.currency_amount), 0) as float)
               from "user" u 
               left outer join deposit d on u.id=d."userId" 
               where d.status='calculated' and now() < d."date" + interval '1 week' * cast(d.ip_wks as int) and u.id=$1
             ) as "currentInvestmentAmount",
             (
-              select cast(coalesce(sum(wh.currency_amount), 0) as int)
+              select cast(coalesce(sum(wh.currency_amount), 0) as float)
               from "user" u 
               left outer join withdraw_history wh on u.id=wh."userId" 
               where u.id=$1
             ) as "totalPayedAmount",
             (
-              select cast(coalesce(sum(d.currency_amount), 0) as int)
+              select cast(coalesce(sum(d.currency_amount), 0) as float)
               from "user" u 
               left outer join deposit d on u.id=d."userId" 
               where d.status='calculated' and u.id=$1
